requirements.txt
pandas
selenium
webdriver-manager
scikit-learn
python-docx
pdfplumber
streamlit
openpyxl